move to input package
